rename trait namespace to concerns
