+++ src/Concerns/HandlesEvents.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hook\Traits;
+namespace Hook\Concerns;
 
 use Hook\Request;
