+++ src/Concerns/MapsEvents.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hook\Traits;
+namespace Hook\Concerns;
 
 trait MapsEvents
 {
